requirements.txt
opencv-python
mss
numpy
pyautogui
Pillow
keyboard
